aider: Updated the loading of the prompts_template in the `agent_actions.py` file to join the list of strings into a single string before formatting it with the input text.
# Aider chat conversation:

ASSISTANT: To update the loading of the prompts_template in the `agent_actions.py` file, we need to modify the way the prompts are loaded from the setup dictionary. Currently, they are loaded as a list of strings, but we need to join them into a single string.

Here is the change:

<source>python
agent_actions.py
<<<<<<< HEAD
        prompt     = self.setup['prompts_templates']["summarize"].format(text=chunk_text)
=======
        prompt     = ''.join(self.setup['prompts_templates']["summarize"]).format(text=chunk_text)
>>>>>>> updated
</source>

<source>python
agent_actions.py
<<<<<<< HEAD
        prompt        = self.setup['prompts_templates']["search"].format(text=inputs[0])
=======
        prompt        = ''.join(self.setup['prompts_templates']["search"]).format(text=inputs[0])
>>>>>>> updated
</source>

This change will join the list of strings into a single string before formatting it with the input text.